Remove redundant in-between files
