+++ frontend/src/component/insights/Insights.tsx
@@ -1,9 +1,12 @@
 import type { FC } from 'react';
 import { styled } from '@mui/material';
 import { InsightsHeader } from './components/InsightsHeader/InsightsHeader.tsx';
-import { InsightsCharts } from './InsightsCharts.tsx';
 import { useUiFlag } from 'hooks/useUiFlag.ts';
 import { LegacyInsights } from './LegacyInsights.tsx';
+import { StyledContainer } from './InsightsCharts.styles.ts';
+import { LifecycleInsights } from './sections/LifecycleInsights.tsx';
+import { PerformanceInsights } from './sections/PerformanceInsights.tsx';
+import { UserInsights } from './sections/UserInsights.tsx';
 
 const StyledWrapper = styled('div')(({ theme }) => ({
     paddingTop: theme.spacing(2),
@@ -13,7 +16,11 @@ const NewInsights: FC = () => {
     return (
         <StyledWrapper>
             <InsightsHeader />
-            <InsightsCharts />
+            <StyledContainer>
+                <LifecycleInsights />
+                <PerformanceInsights />
+                <UserInsights />
+            </StyledContainer>
         </StyledWrapper>
     );
 };
